Fix Access Controls for instances of Subclasses
- Byte needs explicit toInt to access RichInt's String converters
- protected methods are no longer accessible by instances of subsclasses in Scala 3, just within instances of the class and within the subclass itself, but making it package protected opens it up to instances of the subclasses
- scala 3 says charset is doubly defined if it has a def of the same name as the class parameter, so we change the name of the param
- choiceLengthInBits which is a MaybeInt requires explicit toLong to be passed into SpecifiedLengthChoiceParser which takes a long
- Scala 3 no longer considers a self-type as a subclass relationship for access control. So we can't access protected methods, so instead we create a publish getter for the the self-type member and use that instead
- Scala 3 complains of comparing a DebugCommand and a String with "==", so we change the conditional to explicitly check the long and short name
- Scala 3 sometimes uses the wrong superclass/trait as the type of some members, which causes errors when we access methods/members of one superclass/trait that don't exist in the selected superclass/trait, so we explicitly set the types to get rid of these errors
- remove declared isArrayWithAtLeastOneRequiredArrayElement method from ElementBase since it is declared as final in RequiredOptionalMixin
- remove declared groupMembersNotShared method from SequenceGroup since it is declared as final in GroupDefLike
- alignment can be Implicit (if Implicit) or an int (if Prefixed) and is therefore an AnyRef, which cannot be compared to an int with "==" so we cast it to Int since we are in a Prefixed length context
- In Scala 3, backticks can only be used on stable identifiers such as vals, but it doesn't consider instance variables a stable identifier, so we create local vals to get around it
- remove xmlns="" from XML Literals as it results in a weird type mismatch Nil error

DAFFODIL-2975

+++ daffodil-cli/src/main/scala/org/apache/daffodil/cli/Main.scala
@@ -1317,7 +1317,7 @@ class Main(
                         val bitsToDisplay = 8 - bitsAlreadyConsumed
                         val pbp = inStream.inputSource.position() + 1
                         val firstByteBitArray = inStream.getByteArray(bitsToDisplay, finfo)
-                        val fbs = firstByteBitArray(0).toBinaryString
+                        val fbs = firstByteBitArray(0).toInt.toBinaryString
                           .takeRight(8)
                           .reverse
                           .padTo(8, '0')
